Query for fetching location by ID
