+++ src/RootQuery.js
@@ -7,6 +7,7 @@ import AllFlights from './flight/queries/AllFlights';
 import AllAvailableHotels from './hotel/queries/AllAvailableHotels';
 import AvailableHotel from './hotel/queries/AvailableHotel';
 import AllLocations from './location/queries/AllLocations';
+import Location from './location/queries/Location';
 import Booking from './booking/queries/Booking';
 import CurrentUser from './identity/queries/CurrentUser';
 import Hotel from './hotel/queries/Hotel';
@@ -21,6 +22,7 @@ export default new GraphQLObjectType({
     allAvailableHotels: AllAvailableHotels,
     availableHotel: AvailableHotel,
     allLocations: AllLocations,
+    location: Location,
     booking: Booking,
     currentUser: CurrentUser,
     hotel: Hotel,
